+++ tests/Harvest/Harvesters/ItemHarvesterTest.php
@@ -243,6 +243,7 @@ class ItemHarvesterTest extends TestCase
                 '18. storočie, polovica, 1760',
             ],
             'datestamp' => ['2017-08-28T14:00:23.769Z'],
+            'contributor' => 'Čičo, Martin'
         ];
     }
 }